+++ crates/roast/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Update `oas3` dependency to `0.21`.
+
 ## 0.10.0
 
 - Update `oas3` dependency to `0.20`.